Drop --change-ids-file option
Change IDs can be acceptable as a list with --change-ids option or a
file of IDs with --change-ids-file option, but it must be more useful
if both of can be used with the same option.

This update is to enable --change-ids to accept both of a list or file
like as below and drop --change-ids-file useless anymore.

    python3 zuul-client.py -i 888015 874724 other_ids.txt -o res.html